pyproject.toml
[tool.poetry]
name = "polarbear-bot"
version = "0.1.0"
description = "telegram bot by leelhn2345"
authors = ["Nelson Lee <leelhn2345@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.9"
aiogram = "^2.15"
httpx = "^0.19.0"
loguru = "^0.5.3"

[tool.poetry.dev-dependencies]
mypy = "^0.910"
black = "^21.9b0"
pytest = "^6.2.5"
python-dotenv = "^0.19.0"
ipykernel = "^6.4.1"
python-telegram-bot = "^13.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"